+++ src/main/java/com/mojh/daliybudget/common/config/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package com.mojh7.daliybudget.common.config;
+package com.mojh.daliybudget.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
