+++ src/test/java/com/mojh/daliybudget/DaliybudgetApplicationTests.java
@@ -1,4 +1,4 @@
-package com.mojh7.daliybudget;
+package com.mojh.daliybudget;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
